Order Service

+++ account-service/src/main/java/accounts/service/AccountService.java
@@ -4,7 +4,7 @@ import accounts.domain.Account;
 
 public interface AccountService {
 
-    Account findById(int id);
+    Account findById(long id);
 
     Account create(Account account);
